Define scope for database fixtures, and move them to conftest.py
